+++ Advanced_Chrono/app/src/main/java/com/example/advanced_chrono2/model/TimerItem.kt
@@ -3,6 +3,6 @@ package com.example.advanced_chrono2.model
 class TimerItem(
     val id: Int,
     val imageResource: Int,
-    val workoutSeconds: Long,
-    val restSeconds: Long
+    val workoutSeconds: Int,
+    val restSeconds: Int
 )
